Refactored a bit the Daijin code, ignoring the launchers for the assemble/mikado pipeline in coverage analysis. Added tests for the 'is_intersecting' and 'define_graph' sections of superlocus.
