+++ .coveragerc
@@ -5,3 +5,5 @@ omit =
       Mikado/subprograms/util/awk_gtf.py
       Mikado/subprograms/util/class_codes.py
       Mikado/subprograms/util/metrics.py
+      Mikado/daijin/mikado.py
+      Mikado/daijin/assemble.py
